Code commit
Most of the code is done.
Fecthing data from two different websites.
One website is not updating regularly, so it is not used.
The second website works but requires an authentication token that I
currently not able to generate. To obtain it I load the website from the
browser and read that specific value from the cookies then paste it in
the source code.
Sorting and graphs can be made and lots of more things.
Contacted the website developer that I am using now and they said that
they are currently working in this kind of functionality. Yey..
Oh well, I did something at least.
ᕕ( ᐛ )ᕗ