Authentication token for Brawlstats.com
Automatically retrieving the token required to get identified from their
server.
Tried using their javascript files but I could not find ad easy way to
do it.
The solution used is to launch a browser on the website and then reading
the cookie.
The token itself is then stored in a local file together with its expire
date, which I suppose is around 7 hours after its generation, so I set
the expire date after 6 hours.
_へ__(‾◡◝ )>