Brawler rework almost done
The brawler page is almost done.
Set the dimensions for each part of the page to be scalable with the
dimension of the page.
The information box contains power level, trophies and rank of each
brawler.
Added the power level to the brawler and fixed the rank for it.
Testing different fonts now.
୧ʕ•̀ᴥ•́ʔ୨