Rework done
Both reworks needed are done.
The initial player page has the most relevant information about the
player.
The comparisons graphs are now consistent with the rest of the pdf.
Last thing to fix is the player image: it has to be downloaded and then
used appropriately in the pdf file.
(ᴗ˳ᴗ)