Player icon done and brawler power level fix.
Downloaded and associated each player icon to their id.
These images were half in clear on the website and a function has been
added to download them if needed in the future and half hard coded in
the javascripts of the website.
Wrote a few scripts to find, decode and save them locally, then manually
associated to their id (no better way to do this came to my mind
unfortunately).
Fixed a small problem with the power level of brawlers not being
correctly saved on file.
Everything looks correct right now.
In the future, maybe, I could add some more images to show the trophies
and rank of the player more clearly.
Possible extension would be to create a webpage where people can do this
realtime and on their own.
＼＼\(۶•̀ᴗ•́)۶//／／